+++ algorithms/src/main/java/com/leetcode/array/easy/ArrayEasyImpl.java
@@ -100,4 +100,24 @@ public class ArrayEasyImpl implements IArrayEasy{
         }
         return j;
     }
+    /**412. Fizz Buzz*/
+    @Override
+    public List<String> fizzBuzz(int n){
+
+        List<String> list = new ArrayList<>();
+
+        for(int i=1;i<=n;i++)
+        {
+            if(i%3==0 && i%5 ==0)
+                list.add("FizzBuzz");
+            else if(i%3==0)
+                list.add("Fizz");
+            else if(i%5==0)
+                list.add("Buzz");
+            else
+                list.add(i+"");
+        }
+
+        return list;
+    }
 }
