+++ algorithms/src/main/java/com/leetcode/dp/hard/IDPHard.java
@@ -1,0 +1,5 @@
+package com.leetcode.dp.hard;
+/**Dynamic programming hard**/
+
+public interface IDPHard {
+}
