+++ algorithms/src/main/java/com/leetcode/dp/medium/IDPMedium.java
@@ -1,0 +1,7 @@
+package com.leetcode.dp.medium;
+/**Dynamic programming medium*/
+public interface IDPMedium {
+
+    /**5. Longest Palindromic Substring*/
+     String longestPalindrome(String s);
+}
