+++ algorithms/src/main/java/com/leetcode/dp/medium/IDPMediumImpl.java
@@ -1,0 +1,12 @@
+package com.leetcode.dp.medium;
+
+public class IDPMediumImpl implements IDPMedium{
+
+    /**5. Longest Palindromic Substring*/
+    @Override
+     public String longestPalindrome(String s){
+         if(s==null || s.length()<2)
+             return s;
+        return null;//TODO
+     }
+}
