+++ algorithms/src/test/java/com/leetcode/dp/hard/IDPHardTest.java
@@ -1,0 +1,13 @@
+package com.leetcode.dp.hard;
+
+import org.junit.Before;
+
+public class IDPHardTest {
+
+    public IDPHard idph;
+
+    @Before
+    public void init(){
+        idph = new IDPHardImpl();
+    }
+}
