+++ algorithms/src/test/java/com/leetcode/dp/medium/IDPMediumTest.java
@@ -1,0 +1,19 @@
+package com.leetcode.dp.medium;
+
+import org.junit.Before;
+
+/**Dynamic programming medium*/
+public class IDPMediumTest {
+    
+    public IDPMedium idpm;
+    
+    @Before
+    public void init(){
+        idpm = new IDPMediumImpl();
+    }
+
+    /**5. Longest Palindromic Substring*/
+    public void longestPalindromeTest(){
+
+    }
+}
