--- algorithms/src/test/java/com/leetcode/sql/ISqlEasy.java
@@ -1,313 +1,0 @@
-package com.leetcode.sql;
-
-public class ISqlEasy {
-    /**1757. Recyclable and Low Fat Products*/
-     /*
-     Write a solution to find the ids of products that are both low fat and recyclable.
-     products
-     +-------------+---------+
-| Column Name | Type    |
-+-------------+---------+
-| product_id  | int     |
-| low_fats    | enum    |
-| recyclable  | enum    |
-+-------------+---------+
-product_id is the primary key (column with unique values) for this table.
-low_fats is an ENUM (category) of type ('Y', 'N') where 'Y' means this product is low fat and 'N' means it is not.
-recyclable is an ENUM (category) of types ('Y', 'N') where 'Y' means this product is recyclable and 'N' means it is not.
-     * */
-
-
-    /**584. Find Customer Referee*/
-
-/*
-customer
-+-------------+---------+
-| Column Name | Type    |
-+-------------+---------+
-| id          | int     |
-| name        | varchar |
-| referee_id  | int     |
-+-------------+---------+
-In SQL, id is the primary key column for this table.
-Each row of this table indicates the id of a customer, their name, and the id of the customer who referred them.
-
-
-Find the names of the customer that are not referred by the customer with id = 2.
- */
-
-    /**595. Big Countries*/
-    /*
-    world
-    +-------------+---------+
-| Column Name | Type    |
-+-------------+---------+
-| name        | varchar |
-| continent   | varchar |
-| area        | int     |
-| population  | int     |
-| gdp         | bigint  |
-+-------------+---------+
-name is the primary key (column with unique values) for this table.
-Each row of this table gives information about the name of a country, the continent to which it belongs, its area, the population, and its GDP value.
-     A country is big if:
-
-it has an area of at least three million (i.e., 3000000 km2), or
-it has a population of at least twenty-five million (i.e., 25000000).
-Write a solution to find the name, population, and area of the big countries.
-
-Return the result table in any order.
-     */
-
-/** 1148. Article Views I*/
-/*
-views
-+---------------+---------+
-| Column Name   | Type    |
-+---------------+---------+
-| article_id    | int     |
-| author_id     | int     |
-| viewer_id     | int     |
-| view_date     | date    |
-+---------------+---------+
-There is no primary key (column with unique values) for this table, the table may have duplicate rows.
-Each row of this table indicates that some viewer viewed an article (written by some author) on some date.
-Note that equal author_id and viewer_id indicate the same person.
-
-Write a solution to find all the authors that viewed at least one of their own articles.
-Return the result table sorted by id in ascending order.
- */
-
-/**1683. Invalid Tweets */
-/*
-tweets
-+----------------+---------+
-| Column Name    | Type    |
-+----------------+---------+
-| tweet_id       | int     |
-| content        | varchar |
-+----------------+---------+
-tweet_id is the primary key (column with unique values) for this table.
-content consists of alphanumeric characters, '!', or ' ' and no other special characters.
-This table contains all the tweets in a social media app.
-
-
-Write a solution to find the IDs of the invalid tweets. The tweet is invalid if the number of characters used in the content of the tweet is strictly greater than 15.
-Return the result table in any order.
- */
-
-    /**1378. Replace Employee ID With The Unique Identifier*/
-    /*
-    Employees
-    +---------------+---------+
-| Column Name   | Type    |
-+---------------+---------+
-| id            | int     |
-| name          | varchar |
-+---------------+---------+
-id is the primary key (column with unique values) for this table.
-Each row of this table contains the id and the name of an employee in a company.
-
-Table: EmployeeUNI
-
-+---------------+---------+
-| Column Name   | Type    |
-+---------------+---------+
-| id            | int     |
-| unique_id     | int     |
-+---------------+---------+
-(id, unique_id) is the primary key (combination of columns with unique values) for this table.
-Each row of this table contains the id and the corresponding unique id of an employee in the company.
-
-
-Write a solution to show the unique ID of each user, If a user does not have a unique ID replace just show null.
-
-Return the result table in any order.
-     */
-
-    /**1068. Product Sales Analysis I*/
-    /*
-    Table: Sales
-        +-------------+-------+
-| Column Name | Type  |
-+-------------+-------+
-| sale_id     | int   |
-| product_id  | int   |
-| year        | int   |
-| quantity    | int   |
-| price       | int   |
-+-------------+-------+
-(sale_id, year) is the primary key (combination of columns with unique values) of this table.
-product_id is a foreign key (reference column) to Product table.
-Each row of this table shows a sale on the product product_id in a certain year.
-Note that the price is per unit.
-
-
-Table: Product
-
-+--------------+---------+
-| Column Name  | Type    |
-+--------------+---------+
-| product_id   | int     |
-| product_name | varchar |
-+--------------+---------+
-product_id is the primary key (column with unique values) of this table.
-Each row of this table indicates the product name of each product.
-
-Write a solution to report the product_name, year, and price for each sale_id in the Sales table.
-Return the resulting table in any order.
-     */
-
-    /**1581. Customer Who Visited but Did Not Make Any Transactions*/
-    /*
-    Table: Visits
-
-+-------------+---------+
-| Column Name | Type    |
-+-------------+---------+
-| visit_id    | int     |
-| customer_id | int     |
-+-------------+---------+
-visit_id is the column with unique values for this table.
-This table contains information about the customers who visited the mall.
-
-
-Table: Transactions
-
-+----------------+---------+
-| Column Name    | Type    |
-+----------------+---------+
-| transaction_id | int     |
-| visit_id       | int     |
-| amount         | int     |
-+----------------+---------+
-transaction_id is column with unique values for this table.
-This table contains information about the transactions made during the visit_id.
-
-Write a solution to find the IDs of the users who visited without making any transactions and the number of times they made these types of visits.
-Return the result table sorted in any order.
-     */
-
-    /**197. Rising Temperature*/
-    /*
-    Table: Weather
-
-+---------------+---------+
-| Column Name   | Type    |
-+---------------+---------+
-| id            | int     |
-| recordDate    | date    |
-| temperature   | int     |
-+---------------+---------+
-id is the column with unique values for this table.
-There are no different rows with the same recordDate.
-This table contains information about the temperature on a certain day.
-
-Write a solution to find all dates' id with higher temperatures compared to its previous dates (yesterday).
-Return the result table in any order.
-     */
-
-    /**1661. Average Time of Process per Machine*/
-    /*
-    Activity
-
-+----------------+---------+
-| Column Name    | Type    |
-+----------------+---------+
-| machine_id     | int     |
-| process_id     | int     |
-| activity_type  | enum    |
-| timestamp      | float   |
-+----------------+---------+
-The table shows the user activities for a factory website.
-(machine_id, process_id, activity_type) is the primary key (combination of columns with unique values) of this table.
-machine_id is the ID of a machine.
-process_id is the ID of a process running on the machine with ID machine_id.
-activity_type is an ENUM (category) of type ('start', 'end').
-timestamp is a float representing the current time in seconds.
-'start' means the machine starts the process at the given timestamp and 'end' means the machine ends the process at the given timestamp.
-The 'start' timestamp will always be before the 'end' timestamp for every (machine_id, process_id) pair.
-It is guaranteed that each (machine_id, process_id) pair has a 'start' and 'end' timestamp.
-
-
-There is a factory website that has several machines each running the same number of processes. Write a solution to find the average time each machine takes to complete a process.
-
-The time to complete a process is the 'end' timestamp minus the 'start' timestamp. The average time is calculated by the total time to complete every process on the machine divided by the number of processes that were run.
-
-The resulting table should have the machine_id along with the average time as processing_time, which should be rounded to 3 decimal places.
-Return the result table in any order.
-     */
-
-    /**577. Employee Bonus*/
-    /*
-    Table: Employee
-
-+-------------+---------+
-| Column Name | Type    |
-+-------------+---------+
-| empId       | int     |
-| name        | varchar |
-| supervisor  | int     |
-| salary      | int     |
-+-------------+---------+
-empId is the column with unique values for this table.
-Each row of this table indicates the name and the ID of an employee in addition to their salary and the id of their manager.
-
-Table: Bonus
-
-+-------------+------+
-| Column Name | Type |
-+-------------+------+
-| empId       | int  |
-| bonus       | int  |
-+-------------+------+
-empId is the column of unique values for this table.
-empId is a foreign key (reference column) to empId from the Employee table.
-Each row of this table contains the id of an employee and their respective bonus.
-
-Write a solution to report the name and bonus amount of each employee with a bonus less than 1000.
-Return the result table in any order.
-     */
-
-    /**1280. Students and Examinations*/
-    /*
-    Table: Students
-
-+---------------+---------+
-| Column Name   | Type    |
-+---------------+---------+
-| student_id    | int     |
-| student_name  | varchar |
-+---------------+---------+
-student_id is the primary key (column with unique values) for this table.
-Each row of this table contains the ID and the name of one student in the school.
-
-
-Table: Subjects
-
-+--------------+---------+
-| Column Name  | Type    |
-+--------------+---------+
-| subject_name | varchar |
-+--------------+---------+
-subject_name is the primary key (column with unique values) for this table.
-Each row of this table contains the name of one subject in the school.
-
-
-Table: Examinations
-
-+--------------+---------+
-| Column Name  | Type    |
-+--------------+---------+
-| student_id   | int     |
-| subject_name | varchar |
-+--------------+---------+
-There is no primary key (column with unique values) for this table. It may contain duplicates.
-Each student from the Students table takes every course from the Subjects table.
-Each row of this table indicates that a student with ID student_id attended the exam of subject_name.
-
-
-Write a solution to find the number of times each student attended each exam.
-Return the result table ordered by student_id and subject_name.
-     */
-}
