+++ interview-corner/src/main/java/interview/preparation/microservices/principles/principles
@@ -1,0 +1,13 @@
+
+The principles used to design Microservices are as follows:
+
+	1.	Independent & Autonomous Services
+	2.	Scalability
+	3.	Decentralization
+	4.	Resilient Services
+	5.	Real-Time Load Balancing
+	6.	Availability
+	7.	Continuous delivery through DevOps Integration
+	8.	Seamless API Integration and Continuous Monitoring
+	9.	Isolation from Failures //preventing a failure in one microservice from cascading and affecting other services, ensuring the overall system's resilience and reliability
+	10.	Auto -Provisioning //automates the process of deploying and managing resources for microservices, including containers, infrastructure, and even user access
